+++ src/main/java/com/spring/AgendaMedica/servicios/PersonaService.java
@@ -8,7 +8,7 @@ import org.springframework.data.jpa.repository.Query;
 import java.util.List;
 import java.util.Map;
 
-public interface PersonaService extends GenericService<Persona, Long>{
+public interface PersonaService extends GenericService<Persona, Integer>{
 //    @Query(value = "SELECT * FROM persona p JOIN usuarios u ON p.id_persona = u.persona_id_persona WHERE u.username = :username", nativeQuery = true)
 //    public Persona obtenerPersona(String username);
 //
